#164083277 Get specific political party office from database

+++ app/my_api/v2/party.py
@@ -1,6 +1,7 @@
 from flask_restful import Resource, reqparse
 import validators
 import psycopg2
+from flask_jwt_extended import (jwt_required, get_jwt_identity)
 
 # local imports
 from ..database import database
@@ -31,7 +32,7 @@ class CreatePartyV2(Resource):
         help="Logo url needed"
     )
 
-    @classmethod
+    @jwt_required
     def post(self):
         data = CreatePartyV2.parser.parse_args()
         name = ['name']
@@ -51,11 +52,29 @@ class CreatePartyV2(Resource):
                 break
 
         try:
-            cur.execute("INSERT INTO Parties(name, hqaddress, logourl) VALUES(%(name)s, %(hqaddress)s, %(logourl)s);", {
-                'name': data['name'], 'hqaddress': data['hqaddress'], 'logourl': data['logourl']
+            # check if user is admin
+            current_user = get_jwt_identity()
+            cur.execute("SELECT isadmin FROM Users WHERE email = %(email)s;", {
+                'email': current_user
             })
-            connection.commit()
-            return {'Message': 'Party added'}
+            user_exists = cur.fetchone()
+            is_admin = user_exists[0]
+            if is_admin:
+                # check if exists
+                cur.execute("SELECT * FROM Parties WHERE name = %(name)s;", {
+                    'name': data['name']
+                })
+                party_exist = cur.fetchone()
+                if party_exist is None:
+                    cur.execute("INSERT INTO Parties(name, hqaddress, logourl) VALUES(%(name)s, %(hqaddress)s, %(logourl)s);", {
+                        'name': data['name'], 'hqaddress': data['hqaddress'], 'logourl': data['logourl']
+                    })
+                    connection.commit()
+                    return {'Message': 'Party successfully added'}, 200
+                else:
+                    return {'Message': 'Party already exists'}, 409
+            else:
+                return {'Message': 'This panel is for administrators only'}, 403
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
@@ -64,11 +83,12 @@ class CreatePartyV2(Resource):
 
 class GetPartiesV2(Resource):
     """docstring for GetPartiesV2."""
+    @jwt_required
     def get(self):
         try:
             cur.execute("SELECT * FROM Parties;")
             parties = cur.fetchall()
-            return {'Message': parties}
+            return {'Message': parties}, 200
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
@@ -91,18 +111,18 @@ class EditPartyV2(Resource):
         help="Hq address field empty"
     )
 
-    @classmethod
+    @jwt_required
     def get(self, party_id):
         try:
             cur.execute("SELECT * FROM Parties WHERE party_id = %s", [party_id])
             party = cur.fetchall()
-            return {'Message': party}
+            return {'Message': party}, 200
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
             return {'Message': 'current transaction is aborted'}, 500
 
-    @classmethod
+    @jwt_required
     def put(self, party_id):
         data = EditPartyV2.parser.parse_args()
         name = data['name']
@@ -115,19 +135,38 @@ class EditPartyV2(Resource):
             else:
                 break
         try:
-            cur.execute("UPDATE Parties SET name = %s, hqaddress = %s WHERE party_id = %s", (name, hqaddress, party_id))
-            connection.commit()
-            return {'Message': 'The party was updated'}, 200
+            # check for administrator
+            current_user = get_jwt_identity()
+            cur.execute("SELECT isadmin FROM Users WHERE email = %(email)s;", {
+                'email': current_user
+            })
+            user_exists = cur.fetchone()
+            is_admin = user_exists[0]
+            if is_admin:
+                cur.execute("UPDATE Parties SET name = %s, hqaddress = %s WHERE party_id = %s", (name, hqaddress, party_id))
+                connection.commit()
+                return {'Message': 'The party was updated'}, 200
+            else:
+                return {'Message': 'This pannel is for administrators only'}, 403
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
             return {'Message': 'current transaction is aborted'}, 500
 
-    @classmethod
+    @jwt_required
     def delete(self, party_id):
         try:
-            cur.execute("DELETE FROM Parties WHERE party_id = %s;", [party_id])
-            return {'Message': 'Party deleted'}, 200
+            current_user = get_jwt_identity()
+            cur.execute("SELECT isadmin FROM Users WHERE email = %(email)s;", {
+                'email': current_user
+            })
+            user_exists = cur.fetchone()
+            is_admin = user_exists[0]
+            if is_admin:
+                cur.execute("DELETE FROM Parties WHERE party_id = %s;", [party_id])
+                return {'Message': 'Party deleted'}, 200
+            else:
+                return {'Message': 'This pannel is for administrators only'}, 403
         except (Exception, psycopg2.DatabaseError) as error:
             cur.execute("rollback;")
             print(error)
